+++ frontend/src/pages/DubTab.jsx
@@ -1,45 +1,20 @@
-import { Suspense, lazy, useState, useEffect, useCallback, useRef } from 'react';
+import { useState, useEffect, useCallback, useRef } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  Film, Save, UploadCloud, Sparkles, Loader, Square, Users,
-  FileText, Play, DownloadIcon, Volume2, Link2,
-  Languages, ChevronDown, ChevronUp, Wand2, Trash2, Check, Globe, UserSquare2, User, AlertCircle,
-  ShieldCheck,
-} from 'lucide-react';
-// lucide-react exports DownloadIcon as "Download"; alias here to match App.jsx naming.
-import { Download as Download, RotateCcw } from 'lucide-react';
-import SearchableSelect from '../components/SearchableSelect';
-import WaveformTimeline from '../components/WaveformTimeline';
-import CheckpointBanner from '../components/CheckpointBanner';
 import { useAppStore } from '../store';
-import ALL_LANGUAGES from '../languages.json';
-import { POPULAR_LANGS, POPULAR_ISO, PRESETS } from '../utils/constants';
-import { LANG_CODES } from '../utils/languages';
-import { formatTime } from '../utils/format';
 import { API } from '../api/client';
-import { dialectOptionsFor, dialectLabel, dialectMatchesLang } from '../api/dialects';
 import { listTranslationEngines, installTranslationEngine } from '../api/engines';
 import { dubQc } from '../api/dub';
 import toast from 'react-hot-toast';
 import { toastErrorWithReport } from '../utils/errorToast';
-import { Button, Segmented, Badge, Progress } from '../ui';
 import useTimelineOnsets from '../hooks/useTimelineOnsets';
-import GlossaryPanel from '../components/GlossaryPanel';
 import ExportModal from '../components/ExportModal';
-import MultiLangPicker from '../components/MultiLangPicker';
-import DubbingDemo from '../components/DubbingDemo';
-import DubFailureNotice from '../components/dub/DubFailureNotice';
 import DubPipelineStepper from '../components/dub/DubPipelineStepper';
-import PrepOverlay from '../components/dub/PrepOverlay';
-import TranscribeOverlay from '../components/dub/TranscribeOverlay';
-import FooterBtn from '../components/dub/FooterBtn';
+import IdleSkeleton from '../components/dub/IdleSkeleton';
+import DubHeader from '../components/dub/DubHeader';
+import DubLeftColumn from '../components/dub/DubLeftColumn';
+import DubRightColumn from '../components/dub/DubRightColumn';
+import DubFooter from '../components/dub/DubFooter';
 import './DubTab.css';
-
-const DubSegmentTable = lazy(() => import('../components/DubSegmentTable'));
-
-const LazyFallback = () => (
-  <div className="dub-lazy-fallback">Loading…</div>
-);
 
 export default function DubTab(props) {
   const { t, i18n } = useTranslation();
@@ -357,974 +332,18 @@ export default function DubTab(props) {
       )}
       {/* ── Idle: show full editor skeleton with drop zone ── */}
       {showIdleSkeleton && (
-        <div className="dub-col">
-          {/* Header bar */}
-          <div className="dub-head">
-            <div className="label-row dub-head__title">
-                <Film className="label-icon" size={11} />
-                <span className="dub-head__filename">{dubVideoFile ? dubVideoFile.name : t('dub.video_dubbing_studio')}</span>
-              {dubVideoFile && <span className="dub-head__meta">· {(dubVideoFile.size / 1024 / 1024).toFixed(1)} MB</span>}
-              {activeProjectName && activeProjectName !== dubFilename && (
-                <span className="dub-head__project">— {activeProjectName}</span>
-              )}
-            </div>
-            <div className="dub-head__actions">
-              <Button variant="subtle" size="sm" disabled title={t('dub.save')} aria-label={t('dub.save')}><Save size={12} /></Button>
-              <Button variant="ghost" size="sm" disabled title={t('dub.reset')} aria-label={t('dub.reset')}><RotateCcw size={12} /></Button>
-            </div>
-          </div>
-
-          {/* Transcription failure banner — shown in the idle state when a
-              job exists but transcription produced zero segments (or threw).
-              Surfaces the backend error detail and offers one-click retry,
-              which re-runs the ASR stream on the same job without re-uploading. */}
-          {dubError && dubJobId && dubStep === 'idle' && (
-            <div className="dub-footer-banner">
-              <Badge tone="danger">
-                <AlertCircle size={11} /> {dubError}
-              </Badge>
-              <DubFailureNotice failure={dubFailure} />
-              {handleDubRetryTranscribe && (
-                <Button
-                  variant="subtle"
-                  size="sm"
-                  onClick={handleDubRetryTranscribe}
-                  leading={<Sparkles size={10} />}
-                >
-                  {t('dub.retry_transcription')}
-                </Button>
-              )}
-              {handleDubImportSrt && (
-                <label
-                  htmlFor="srt-import-banner-input"
-                  className="dub-idle-upload-label"
-                  title={t('dub.import_srt')}
-                  style={{ cursor: 'pointer' }}
-                >
-                  <FileText size={11} /> {t('dub.import_srt_alt')}
-                  <input
-                    id="srt-import-banner-input"
-                    type="file"
-                    accept=".srt,text/srt,text/plain"
-                    hidden
-                    onChange={(e) => {
-                      const f = e.target.files?.[0];
-                      if (f) handleDubImportSrt(f);
-                      e.target.value = '';
-                    }}
-                  />
-                </label>
-              )}
-            </div>
-          )}
-
-          {/* SPLIT LAYOUT skeleton */}
-          <div className={`dub-split-grid ${dubVideoFile ? 'dub-split-2' : 'dub-split-1'}`}>
-            {/* LEFT */}
-            <div className="studio-panel dub-panel-col">
-              {dubVideoFile ? (
-                <>
-                  <WaveformTimeline
-                    audioSrc={dubLocalBlobUrl?.audioUrl}
-                    videoSrc={dubLocalBlobUrl?.videoUrl}
-                    segments={[]}
-                    disabled={true}
-                    overlayContent={
-                      dubStep === 'uploading' ? (
-                        <PrepOverlay stage={dubPrepStage} progress={dubPrepProgress} onAbort={handleDubAbort} />
-                      ) : dubStep === 'transcribing' ? (
-                        <TranscribeOverlay
-                          elapsed={transcribeElapsed}
-                          duration={dubDuration}
-                          onAbort={handleDubAbort}
-                        />
-                      ) : null
-                    }
-                  />
-                  <div className="dub-change-row">
-                    <label htmlFor="video-upload" className="dub-idle-upload-label">
-                      <Film size={13} /> {t('dub.change_file')}
-                    </label>
-                    {dubJobId && handleDubImportSrt && (
-                      <label
-                        htmlFor="srt-import-input"
-                        className="dub-idle-upload-label"
-                        title={t('dub.import_srt')}
-                        style={{ cursor: 'pointer' }}
-                      >
-                        <FileText size={13} /> {t('dub.import_srt')}
-                        <input
-                          id="srt-import-input"
-                          type="file"
-                          accept=".srt,text/srt,text/plain"
-                          hidden
-                          onChange={(e) => {
-                            const f = e.target.files?.[0];
-                            if (f) handleDubImportSrt(f);
-                            e.target.value = '';
-                          }}
-                        />
-                      </label>
-                    )}
-                    <label className="dub-speakers-hint" title={t('dub.num_speakers_help')}>
-                      <Users size={13} /> {t('dub.num_speakers_label')}
-                      <input
-                        type="number"
-                        min={1}
-                        max={20}
-                        step={1}
-                        className="dub-speakers-input"
-                        placeholder={t('dub.num_speakers_auto')}
-                        value={dubNumSpeakers ?? ''}
-                        disabled={dubStep === 'uploading' || dubStep === 'transcribing'}
-                        onChange={(e) => {
-                          const v = parseInt(e.target.value, 10);
-                          setDubNumSpeakers(Number.isFinite(v) && v > 0 ? Math.min(v, 20) : null);
-                        }}
-                      />
-                    </label>
-                    <button className="btn-primary dub-change-row__cta"
-                      onClick={handleDubUpload}
-                      disabled={dubStep === 'uploading' || dubStep === 'transcribing'}>
-                      {dubStep === 'uploading' || dubStep === 'transcribing'
-                        ? <><Loader className="spinner" size={14} /> {t('common.loading')}</>
-                        : <><Sparkles size={14} /> {t('dub.upload_transcribe')}</>}
-                    </button>
-                  </div>
-                </>
-              ) : dubStep === 'uploading' ? (
-                <PrepOverlay stage={dubPrepStage} progress={dubPrepProgress} onAbort={handleDubAbort} large />
-              ) : (
-                <>
-                  {!demoDismissed && (
-                    <DubbingDemo onDismiss={dismissDubDemo} />
-                  )}
-                  <label htmlFor="video-upload" className="dub-idle-drop"
-                  onDragOver={e => { e.preventDefault(); e.currentTarget.classList.add('is-dragging'); }}
-                  onDragLeave={e => { e.currentTarget.classList.remove('is-dragging'); }}
-                  onDrop={e => {
-                    e.preventDefault();
-                    e.currentTarget.classList.remove('is-dragging');
-                    const file = e.dataTransfer.files[0];
-                    if (file && (file.type.startsWith('video/') || file.type.startsWith('audio/') || /\.(mp3|wav|flac|m4a|aac|ogg|opus|wma)$/i.test(file.name))) {
-                      setDubVideoFile(file);
-                      // #119: an audio file → audio-only dubbing (skip video work, output audio).
-                      setDubInputType(file.type.startsWith('audio/') || /\.(mp3|wav|flac|m4a|aac|ogg|opus|wma)$/i.test(file.name) ? 'audio' : 'video');
-                      setDubStep('idle');
-                      fileToMediaUrl(file, null).then(urls => setDubLocalBlobUrl(urls));
-                    }
-                  }}>
-                  <div className="dub-idle-drop__puck">
-                    <UploadCloud color="#d3869b" size={28} />
-                  </div>
-                  <div className="dub-idle-drop__lines">
-                    <div className="dub-idle-drop__title">{t('dub.drop_here')}</div>
-                    <div className="dub-idle-drop__sub">{t('dub.supported_formats')}</div>
-                  </div>
-                  <div
-                    className="dub-ingest-row"
-                    onClick={e => e.preventDefault()}
-                  >
-                    <Link2 size={13} color="#a89984" />
-                    <input
-                      type="text"
-                      placeholder={t('dub.paste_url')}
-                      value={ingestUrl}
-                      onChange={e => setIngestUrl(e.target.value)}
-                      onClick={e => { e.preventDefault(); e.stopPropagation(); }}
-                      onKeyDown={e => { if (e.key === 'Enter') { e.preventDefault(); e.stopPropagation(); onIngestUrl(); } }}
-                      className="dub-ingest-row__input"
-                    />
-                    <button
-                      type="button"
-                      onClick={e => { e.preventDefault(); e.stopPropagation(); onIngestUrl(); }}
-                      disabled={!ingestUrl.trim()}
-                      className={`dub-ingest-row__cta ${ingestUrl.trim() ? 'is-ready' : ''}`}
-                    >
-                      {t('dub.ingest')}
-                    </button>
-                  </div>
-                  <label
-                    className="dub-ingest-sub-opt"
-                    title={t('dub.pull_captions_title')}
-                    onClick={e => { e.stopPropagation(); }}
-                  >
-                    <input
-                      type="checkbox"
-                      checked={fetchYtSubs}
-                      onChange={e => setFetchYtSubs(e.target.checked)}
-                      onClick={e => e.stopPropagation()}
-                    />
-                    <span>{t('dub.pull_captions')}</span>
-                  </label>
-                </label>
-
-                {/* One decision up front: the target language. Everything else
-                    (speakers, style) hides behind Advanced — ElevenLabs-style
-                    flow, OmniVoice chrome. The pick pre-seeds the editor. */}
-                <div className="dub-landing-opts">
-                  <label className="dub-landing-opts__lang">
-                    <Globe size={13} />
-                    <span className="dub-landing-opts__label">{t('dub.target_language', { defaultValue: 'Dub into' })}</span>
-                    <select
-                      className="input-base input-base--xs"
-                      value={dubLangCode}
-                      onChange={(e) => {
-                        const lc = LANG_CODES.find(l => l.code === e.target.value);
-                        setDubLangCode(e.target.value);
-                        if (lc) setDubLang(lc.label);
-                      }}
-                    >
-                      {LANG_CODES.map(lc => (
-                        <option key={lc.code} value={lc.code}>{lc.label} — {lc.code}</option>
-                      ))}
-                    </select>
-                  </label>
-                  <button
-                    type="button"
-                    className="dub-landing-opts__adv"
-                    onClick={() => setLandingAdvOpen(o => !o)}
-                    aria-expanded={landingAdvOpen}
-                  >
-                    {t('dub.advanced', { defaultValue: 'Advanced' })}
-                    {landingAdvOpen ? <ChevronUp size={11} /> : <ChevronDown size={11} />}
-                  </button>
-                </div>
-                {landingAdvOpen && (
-                  <div className="dub-landing-adv">
-                    <label className="dub-landing-adv__field" title={t('dub.num_speakers_help')}>
-                      <Users size={12} /> {t('dub.num_speakers_label')}
-                      <input
-                        type="number" min={1} max={20} step={1}
-                        className="input-base input-base--xs dub-speakers-input"
-                        placeholder={t('dub.num_speakers_auto')}
-                        value={dubNumSpeakers ?? ''}
-                        onChange={(e) => {
-                          const v = parseInt(e.target.value, 10);
-                          setDubNumSpeakers(Number.isFinite(v) && v > 0 ? Math.min(v, 20) : null);
-                        }}
-                      />
-                    </label>
-                    <label className="dub-landing-adv__field dub-landing-adv__field--grow">
-                      <UserSquare2 size={12} /> {t('dub.style')}
-                      <input
-                        type="text"
-                        className="input-base input-base--xs"
-                        placeholder={t('dub.style_placeholder')}
-                        value={dubInstruct}
-                        onChange={(e) => setDubInstruct(e.target.value)}
-                      />
-                    </label>
-                  </div>
-                )}
-                </>
-              )}
-
-              <input type="file" accept="video/*,audio/*,.mp3,.wav,.m4a,.aac,.flac,.ogg,.opus,.wma" id="video-upload" className="dub-hidden-file"
-                onChange={e => {
-                  const file = e.target.files[0];
-                  if (!file) return;
-                  setDubVideoFile(file);
-                  // #119: an audio file → audio-only dubbing (skip video work, output audio).
-                  setDubInputType(file.type.startsWith('audio/') || /\.(mp3|wav|flac|m4a|aac|ogg|opus|wma)$/i.test(file.name) ? 'audio' : 'video');
-                  setDubStep('idle');
-                  setDubLocalBlobUrl(prev => { fileToMediaUrl(file, prev).then(urls => setDubLocalBlobUrl(urls)); return prev; });
-                }} />
-
-              {dubVideoFile && (
-                <div className="dub-cast dub-cast--muted">
-                  <div className="dub-cast__row">
-                    <span className="dub-cast__kicker">{t('dub.cast')}</span>
-                    <span className="dub-cast__label">{t('dub.speaker', { n: 1 })}</span>
-                    <span className="dub-cast--muted__chip">{t('dub.default')}</span>
-                  </div>
-                </div>
-              )}
-            </div>
-
-            {/* RIGHT: Ghost settings + segment table (only when video loaded) */}
-            {dubVideoFile ? (
-            <div className="studio-panel dub-panel-col">
-              <div className="dub-skel-settings">
-                <div className="dub-skel-field">
-                  <div className="label-row"><Globe className="label-icon" size={9} /> {t('dub.language')}</div>
-                  <select className="input-base input-base--xs" disabled>
-                    <option>{t('dub.auto')}</option>
-                  </select>
-                </div>
-                <div className="dub-skel-field--sm">
-                  <div className="label-row">{t('dub.iso_code')}</div>
-                  <select className="input-base input-base--xs" disabled>
-                    <option>en — {t('dub.original_audio')}</option>
-                  </select>
-                </div>
-                <div className="dub-skel-field">
-                  <div className="label-row"><UserSquare2 className="label-icon" size={9} /> {t('dub.style')}</div>
-                  <input className="input-base input-base--xs" disabled placeholder={t('dub.style_placeholder')} />
-                </div>
-                <button disabled className="dub-skel-translate-btn">
-                  <Languages size={10} /> {t('dub.translate_all')}
-                </button>
-              </div>
-              <div className="dub-skel-transcript-toggle">
-                <div className="override-toggle dub-skel-transcript-toggle__inner">
-                  <span><FileText size={10} className="dub-inline-icon" /> {t('dub.transcript')}</span>
-                  <ChevronDown size={10} />
-                </div>
-              </div>
-              <div className="segment-table dub-skel-table">
-                <div className="segment-header">
-                  <span className="dub-skel-header-time">{t('dub.time_col')}</span>
-                  <span className="dub-skel-header-spkr">{t('dub.spkr_col')}</span>
-                  <span className="dub-skel-header-text">{t('dub.text_col')}</span>
-                  <span className="dub-skel-header-voice">{t('dub.voice_col')}</span>
-                  <span className="dub-skel-header-acts"></span>
-                </div>
-                {[1, 2, 3, 4, 5, 6].map(i => (
-                  <div key={i} className="segment-row dub-skel-row" style={{ opacity: 0.5 + (0.07 * (6 - i)) }}>
-                    <span className="dub-skel-cell-time dub-skel-bar" />
-                    <span className="dub-skel-cell-spkr dub-skel-bar" />
-                    <div className="dub-skel-cell-text dub-skel-bar" />
-                    <span className="dub-skel-cell-voice dub-skel-bar" />
-                    <div className="dub-skel-cell-acts">
-                      <span className="segment-del dub-skel-cell-acts__icon"><Trash2 size={9} /></span>
-                    </div>
-                  </div>
-                ))}
-                <div className="dub-skel-hint">{t('dub.transcript_after_extract', { defaultValue: 'Transcript appears after extraction.' })}</div>
-              </div>
-            </div>
-            ) : null}
-          </div>
-
-          {/* Ghost footer — only once a file is in play; the bare landing stays
-              clean. Generate is the lone primary, exports demoted to one menu. */}
-          {dubVideoFile && (
-            <div className="studio-panel dub-ghost-footer">
-              <div className="dub-skel-gen-row">
-                <button className="btn-primary dub-skel-gen-btn" disabled>
-                  <Play size={11} /> {t('dub.generate_dub')}
-                </button>
-                <button className="dub-skel-gen-btn dub-skel-gen-btn--secondary" disabled>
-                  <Download size={11} /> {t('dub.export_btn', { defaultValue: 'Export' })} <ChevronDown size={10} />
-                </button>
-              </div>
-            </div>
-          )}
-        </div>
+        <IdleSkeleton t={t} dubVideoFile={dubVideoFile} activeProjectName={activeProjectName} dubFilename={dubFilename} dubError={dubError} dubJobId={dubJobId} dubStep={dubStep} dubFailure={dubFailure} handleDubRetryTranscribe={handleDubRetryTranscribe} handleDubImportSrt={handleDubImportSrt} dubLocalBlobUrl={dubLocalBlobUrl} dubPrepStage={dubPrepStage} dubPrepProgress={dubPrepProgress} handleDubAbort={handleDubAbort} transcribeElapsed={transcribeElapsed} dubDuration={dubDuration} dubNumSpeakers={dubNumSpeakers} setDubNumSpeakers={setDubNumSpeakers} handleDubUpload={handleDubUpload} demoDismissed={demoDismissed} dismissDubDemo={dismissDubDemo} setDubVideoFile={setDubVideoFile} setDubInputType={setDubInputType} setDubStep={setDubStep} fileToMediaUrl={fileToMediaUrl} setDubLocalBlobUrl={setDubLocalBlobUrl} ingestUrl={ingestUrl} setIngestUrl={setIngestUrl} onIngestUrl={onIngestUrl} fetchYtSubs={fetchYtSubs} setFetchYtSubs={setFetchYtSubs} dubLangCode={dubLangCode} setDubLangCode={setDubLangCode} setDubLang={setDubLang} landingAdvOpen={landingAdvOpen} setLandingAdvOpen={setLandingAdvOpen} dubInstruct={dubInstruct} setDubInstruct={setDubInstruct} />
       )}
 
       {/* ── After transcription: side-by-side editor ── */}
       {dubJobId && (dubStep === 'editing' || dubStep === 'generating' || dubStep === 'done') && (
         <div className="dub-col">
-          <div className="dub-head">
-            <div className="label-row dub-head__title">
-              <FileText className="label-icon" size={11} />
-              <span className="dub-head__filename">{dubFilename}</span>
-              <span className="dub-head__meta">· {formatTime(dubDuration)} · {dubSegments.length} {t('dub.segs')}</span>
-              {activeProjectName && activeProjectName !== dubFilename && (
-                <span className="dub-head__project">— {activeProjectName}</span>
-              )}
-            </div>
-            <div className="dub-head__actions">
-              {/* Icon-only secondary actions (tooltips carry the labels);
-                  Generate Dub keeps its label as the primary verb. */}
-              <Button variant="subtle" size="sm" onClick={saveProject}
-                title={t('dub.save')} aria-label={t('dub.save')}><Save size={12} /></Button>
-              <Button variant="danger" size="sm" onClick={resetDub}
-                title={t('dub.reset')} aria-label={t('dub.reset')}><RotateCcw size={12} /></Button>
-              {/* Primary actions live on the header bar (compact) — moved up from the footer. */}
-              <div className="dub-head__primary">
-                {dubStep === 'stopping' ? (
-                  <FooterBtn sm tone="stopping" disabled icon={<Loader className="spinner" size={9} />} label={t('dub.stopping')} />
-                ) : dubStep === 'generating' ? (
-                  <FooterBtn sm tone="danger" onClick={handleDubStop} icon={<Square size={9} />}
-                    label={t('dub.stop_progress', { current: dubProgress.current, total: dubProgress.total })} />
-                ) : (
-                  <>
-                    <FooterBtn sm tone={dubSegments.length ? 'pink' : 'idle'} onClick={onGenerateClick}
-                      disabled={!dubSegments.length} icon={<Play size={11} />}
-                      label={multiLangMode && multiLangs.length > 1
-                        ? t('dub.generate_dub_multi', { count: multiLangs.length, defaultValue: 'Generate {{count}} dubs' })
-                        : t('dub.generate_dub')} />
-                    {dubStep === 'done' && incrementalPlan && incrementalPlan.stale?.length > 0 && (
-                      <FooterBtn sm tone="pink"
-                        onClick={() => handleDubGenerate({ regenOnly: incrementalPlan.stale, preview: true })}
-                        icon={<Play size={11} />}
-                        label={t('dub.regen_changed', { count: incrementalPlan.stale.length })} />
-                    )}
-                  </>
-                )}
-                {dubStep === 'done' && (
-                  <FooterBtn sm tone="idle"
-                    disabled={qcRunning || !dubSegments.length}
-                    onClick={handleDubQc}
-                    icon={qcRunning ? <Loader className="spinner" size={11} /> : <ShieldCheck size={11} />}
-                    title={t('dub.qc_btn', { defaultValue: 'Verify dub timing (second-pass check)' })}
-                    aria-label={t('dub.qc_btn', { defaultValue: 'Verify dub timing (second-pass check)' })} />
-                )}
-                <FooterBtn sm tone={dubStep === 'done' ? 'green' : 'idle'}
-                  disabled={dubStep !== 'done' && !dubSegments.length}
-                  onClick={() => setExportOpen(true)}
-                  icon={<Download size={12} />}
-                  title={t('dub.export_btn')} aria-label={t('dub.export_btn')} />
-              </div>
-            </div>
-          </div>
-
+          <DubHeader t={t} dubFilename={dubFilename} dubDuration={dubDuration} dubSegments={dubSegments} activeProjectName={activeProjectName} saveProject={saveProject} resetDub={resetDub} dubStep={dubStep} handleDubStop={handleDubStop} dubProgress={dubProgress} onGenerateClick={onGenerateClick} multiLangMode={multiLangMode} multiLangs={multiLangs} incrementalPlan={incrementalPlan} handleDubGenerate={handleDubGenerate} qcRunning={qcRunning} handleDubQc={handleDubQc} setExportOpen={setExportOpen} />
           <div className="dub-split-grid dub-split-2">
-            {/* LEFT: Waveform + Video */}
-            <div className="studio-panel dub-panel-col">
-              {hasDubbedTrack && (
-                <div className="dub-lang-switch" role="radiogroup" aria-label={t('dub.preview_language', { defaultValue: 'Preview language' })}>
-                  <button
-                    type="button"
-                    role="radio"
-                    aria-checked={previewMode === 'original'}
-                    className={`dub-lang-pill ${previewMode === 'original' ? 'is-active' : ''}`}
-                    onClick={() => setPreviewMode('original')}
-                  >
-                    {t('dub.original_audio')}
-                  </button>
-                  {dubTracks.map(code => {
-                    const label = LANG_CODES.find(lc => lc.code === code)?.label || code.toUpperCase();
-                    return (
-                      <button
-                        key={code}
-                        type="button"
-                        role="radio"
-                        aria-checked={previewMode === code}
-                        className={`dub-lang-pill ${previewMode === code ? 'is-active' : ''}`}
-                        onClick={() => setPreviewMode(code)}
-                      >
-                        {label}
-                      </button>
-                    );
-                  })}
-                </div>
-              )}
-              <WaveformTimeline
-                key={videoSrc}
-                ref={waveformRef}
-                audioSrc={`${API}/dub/audio/${dubJobId}`}
-                videoSrc={videoSrc}
-                segments={dubSegments}
-                onsets={timelineOnsets}
-                selectedSegId={timelineSelSegId}
-                onSelectSeg={setTimelineSelSegId}
-                incrementalPlan={incrementalPlan}
-                onSegmentCommit={segmentMoveResize}
-                onSegmentDelete={segmentDelete}
-                onPreviewSegment={onTimelinePreviewSegment}
-                disabled={dubStep === 'generating' || dubStep === 'stopping'}
-                overlayContent={(dubStep === 'generating' || dubStep === 'stopping') ? (
-                  <div className="dub-gen-overlay">
-                    <div className="dub-gen-overlay__head">
-                      {dubStep === 'stopping' ? <Loader className="spinner" size={14} color="#a89984" /> : <Sparkles className="spinner" size={14} color="#d3869b" />}
-                      <span className={`dub-gen-overlay__title ${dubStep === 'stopping' ? 'is-stopping' : ''}`}>
-                        {dubStep === 'stopping' ? t('dub.stopping') : t('dub.generate_dub') + ` ${dubProgress.current}/${dubProgress.total}…`}
-                      </span>
-                    </div>
-                    {dubStep === 'generating' && (
-                      <>
-                        <div className="dub-gen-overlay__stats">
-                          <span>⏱ {fmtDur(genElapsed)} {t('dub.elapsed')}</span>
-                          {genRemaining !== null && <span>~{fmtDur(genRemaining)} {t('dub.remaining')}</span>}
-                        </div>
-                        <div className="dub-gen-overlay__bar">
-                          <Progress
-                            value={dubProgress.total ? (dubProgress.current / dubProgress.total) * 100 : 0}
-                            tone="brand"
-                            size="sm"
-                          />
-                        </div>
-                        {dubProgress.text && <span className="dub-gen-overlay__text">{dubProgress.text}</span>}
-                      </>
-                    )}
-                  </div>
-                ) : null}
-              />
-
-              {/* Cast — per-speaker voice assignment. When the auto-clone
-                  extractor found a usable passage per speaker (≥5s from the
-                  isolated vocals), that option becomes first-class in the
-                  dropdown. It's also pre-selected on the segments so "new
-                  language = same speaker's voice" works by default. */}
-              {dubSegments.some(s => s.speaker_id) && (
-                <div className="dub-cast">
-                  <div className="dub-cast__row">
-                    <span className="dub-cast__kicker" title={t('dub.cast_title')}>{t('dub.cast')}</span>
-                    {[...new Set(dubSegments.map(s => s.speaker_id).filter(Boolean))].map(spk => {
-                      const autoId = `auto:${(spk || '').toLowerCase().replace(/\s+/g, '_')}`;
-                      const clone = speakerClones[spk];
-                      return (
-                        <div key={spk} className="dub-cast__pair">
-                          <span className="dub-cast__label">{spk}:</span>
-                          <select className="input-base dub-cast__select"
-                            value={dubSegments.find(s => s.speaker_id === spk)?.profile_id || ''}
-                            onChange={e => {
-                              const val = e.target.value;
-                              setDubSegments(dubSegments.map(s => s.speaker_id === spk ? { ...s, profile_id: val } : s));
-                            }}>
-                            {clone && (
-                              <option value={autoId}>{t('dub.from_video', { duration: clone.duration.toFixed(1) })}</option>
-                            )}
-                            <option value="">{t('dub.default')}</option>
-                            {profiles.length > 0 && (
-                              <optgroup label={t('dub.clone_profiles')}>
-                                {profiles.map(p => <option key={p.id} value={p.id}>{p.name}</option>)}
-                              </optgroup>
-                            )}
-                            {PRESETS.length > 0 && (
-                              <optgroup label={t('dub.design_presets')}>
-                                {PRESETS.map(p => <option key={p.id} value={`preset:${p.id}`}>{p.name}</option>)}
-                              </optgroup>
-                            )}
-                          </select>
-                        </div>
-                      );
-                    })}
-                  </div>
-                </div>
-              )}
-
-              {/* Translation settings — collapsed or expanded */}
-              {!settingsOpen && (
-                <div className="dub-settings-summary">
-                  <button
-                    type="button"
-                    className="dub-settings-summary__trigger"
-                    onClick={() => setSettingsOpen(true)}
-                    title={t('dub.edit_settings')}
-                  >
-                    <ChevronDown size={10} />
-                    <span><strong>{dubLang}</strong> · {dubLangCode} · {translateQuality} · <span style={{ color: activeEngineUnavailable ? '#fb4934' : '#b8bb26' }}>●</span> {translateProvider}</span>
-                    {dubInstruct && <span className="dub-settings-summary__style">{t('dub.style_label_prefix')}{dubInstruct}</span>}
-                  </button>
-                  <Button
-                    variant="subtle" size="sm"
-                    onClick={handleTranslateAll}
-                    disabled={isTranslating || !dubSegments.length}
-                    loading={isTranslating}
-                    leading={!isTranslating && <Languages size={10} />}
-                  >
-                    {isTranslating ? t('dub.translating') : hasAnyTranslation ? t('dub.retranslate') : t('dub.translate_all')}
-                  </Button>
-                  <Button
-                    variant="subtle" size="sm"
-                    onClick={handleCleanupSegments}
-                    disabled={!dubSegments.length || !dubJobId}
-                    title={t('dub.clean_up_title')}
-                    leading={<Wand2 size={10} />}
-                  >
-                    {t('dub.clean_up')}
-                  </Button>
-                </div>
-              )}
-              {settingsOpen && (
-              <div className="dub-settings-bar">
-                <div className="dub-settings-bar__fields">
-                  <button
-                    type="button"
-                    className="dub-settings-summary__trigger dub-settings-close"
-                    onClick={() => setSettingsOpen(false)}
-                    title={t('dub.collapse_settings')}
-                  >
-                    <ChevronUp size={10} />
-                  </button>
-                  <div className="dub-settings-field dub-settings-field--lang">
-                    <div className="label-row"><Globe className="label-icon" size={9} /> {t('dub.language')}</div>
-                    <select
-                      className="input-base dub-cast__select"
-                      value={dubLang}
-                      onChange={(e) => {
-                        const lang = e.target.value;
-                        setDubLang(lang);
-                        const match = LANG_CODES.find(lc => lc.label.toLowerCase() === lang.toLowerCase());
-                        if (match) {
-                          setDubLangCode(match.code);
-                          // #280: a dialect belongs to one language — clear it
-                          // whenever the new target doesn't match.
-                          if (!dialectMatchesLang(dubDialect, match.code)) setDubDialect('');
-                        }
-                      }}
-                    >
-                      <optgroup label={t('dub.popular')}>
-                        {POPULAR_LANGS.map(l => <option key={`p-${l}`} value={l}>{l}</option>)}
-                      </optgroup>
-                      <optgroup label={t('dub.all_languages')}>
-                        {ALL_LANGUAGES
-                          .filter(l => !POPULAR_LANGS.includes(l))
-                          .map(l => <option key={l} value={l}>{l}</option>)}
-                      </optgroup>
-                    </select>
-                  </div>
-                  <div className="dub-settings-field dub-settings-field--iso">
-                    <div className="label-row">{t('dub.iso_code')}</div>
-                    <select
-                      className="input-base dub-cast__select"
-                      value={dubLangCode}
-                      onChange={(e) => {
-                        const code = e.target.value;
-                        setDubLangCode(code);
-                        if (!dialectMatchesLang(dubDialect, code)) setDubDialect('');
-                      }}
-                    >
-                      {LANG_CODES.map(lc => (
-                        <option key={lc.code} value={lc.code}>{lc.code} — {lc.label}</option>
-                      ))}
-                    </select>
-                  </div>
-                  {/* #280: regional dialect / vocabulary. Only rendered for
-                      languages with curated variants; region names come from
-                      Intl.DisplayNames so they localize with the UI for free. */}
-                  {dialectOptionsFor(dubLangCode).length > 0 && (
-                    <div className="dub-settings-field dub-settings-field--dialect">
-                      <div className="label-row" title={t('dub.dialect_title')}>{t('dub.dialect_label')}</div>
-                      <select
-                        className="input-base dub-cast__select"
-                        value={dialectMatchesLang(dubDialect, dubLangCode) ? dubDialect : ''}
-                        onChange={(e) => setDubDialect(e.target.value)}
-                      >
-                        <option value="">{t('dub.dialect_default')}</option>
-                        {dialectOptionsFor(dubLangCode).map(d => (
-                          <option key={d} value={d}>{dialectLabel(d, i18n.language)}</option>
-                        ))}
-                      </select>
-                    </div>
-                  )}
-                  <div className="dub-settings-field dub-settings-field--engine">
-                    <div className="label-row">
-                      {t('dub.engine_label')}
-                      {activeEngineUnavailable && !enginesSandboxed && (
-                        <button
-                          type="button"
-                          className="dub-engine-install-chip"
-                          onClick={() => handleInstallEngine(translateProvider)}
-                          disabled={engineInstalling === translateProvider}
-                          title={t('dub.install_engine')}
-                        >
-                          {engineInstalling === translateProvider ? t('dub.installing_engine') : `+ install ${activeEngineEntry?.pip_package || ''}`}
-                        </button>
-                      )}
-                      {activeEngineUnavailable && enginesSandboxed && (
-                        <span className="dub-engine-install-chip dub-engine-install-chip--disabled" title={t('dub.install_disabled_title')}>
-                          {t('dub.needs_dev_install')}
-                        </span>
-                      )}
-                    </div>
-                    <select className="input-base dub-engine-select" value={translateProvider} onChange={e => setTranslateProvider(e.target.value)}>
-                      {(engines.length ? engines : []).map(p => (
-                        <option key={p.id} value={p.id}>
-                          {p.installed ? p.display_name : `${p.display_name}${t('dub.needs_install_suffix')}`}
-                        </option>
-                      ))}
-                    </select>
-                  </div>
-                  <div className="dub-settings-field dub-settings-field--quality">
-                    <div className="label-row" title={t('dub.quality_title')}>{t('dub.quality_label')}</div>
-                    <Segmented
-                      size="sm"
-                      value={translateQuality}
-                      onChange={(v) => {
-                        // #372: picking Cinematic with no LLM configured used to
-                        // bounce the user between two warnings forever. Block the
-                        // pick at the source and point at the actual fix.
-                        if (v === 'cinematic' && llmEndpoint && !llmEndpoint.available) {
-                          toast(t('dub.cinematic_needs_llm_hint', { defaultValue: 'Cinematic needs an LLM. Configure one in Settings → Credentials → LLM endpoint (Ollama runs locally, no key needed).' }), { icon: 'ℹ️', duration: 8000 });
-                          return;
-                        }
-                        setTranslateQuality(v);
-                      }}
-                      items={[
-                        { value: 'fast',      label: t('dub.fast_quality') },
-                        { value: 'cinematic', label: t('dub.cinematic_quality') },
-                      ]}
-                    />
-                  </div>
-                  <div className="dub-settings-field dub-settings-field--style">
-                    <div className="label-row"><UserSquare2 className="label-icon" size={9} /> {t('dub.style')} <span className="dub-settings-field__hint">{t('dub.optional')}</span></div>
-                    <input className="input-base input-base--xs" placeholder={t('dub.style_placeholder')} value={dubInstruct} onChange={e => setDubInstruct(e.target.value)} />
-                  </div>
-                  <div className="dub-settings-field dub-settings-field--multi">
-                    <label className="dub-multi-toggle">
-                      <input
-                        type="checkbox"
-                        checked={multiLangMode}
-                        onChange={e => setMultiLangMode(e.target.checked)}
-                      />
-                      <span>{t('dub.multi_lang')}</span>
-                    </label>
-                    {multiLangMode && (
-                      <MultiLangPicker
-                        selected={multiLangs}
-                        onChange={setMultiLangs}
-                        disabled={dubStep === 'generating'}
-                      />
-                    )}
-                  </div>
-                </div>
-                <div className="dub-settings-bar__actions">
-                  <Button
-                    variant="subtle" size="sm"
-                    onClick={() => editSegments(dubSegments.map(s => ({ ...s, text: s.text_original || s.text, translate_error: undefined })))}
-                    disabled={!dubSegments.some(s => s.text_original && s.text_original !== s.text)}
-                    title={t('dub.restore_title')}
-                  >
-                    {t('dub.restore')}
-                  </Button>
-                  <Button
-                    variant="subtle" size="sm"
-                    onClick={handleCleanupSegments}
-                    disabled={!dubSegments.length || !dubJobId}
-                    title={t('dub.clean_up_title')}
-                    leading={<Wand2 size={10} />}
-                  >
-                    {t('dub.clean_up')}
-                  </Button>
-                  <Button
-                    variant="primary" size="sm"
-                    onClick={handleTranslateAll}
-                    disabled={isTranslating || !dubSegments.length}
-                    loading={isTranslating}
-                    leading={!isTranslating && <Languages size={10} />}
-                  >
-                    {isTranslating ? t('dub.translating') : t('dub.translate_all')}
-                  </Button>
-                </div>
-              </div>
-              )}
-            </div>
-
-            {/* RIGHT: Segment Table */}
-            <div className="studio-panel dub-panel-col">
-
-              {/* Output options + timing — moved to the top of the right section. */}
-              <div className="dub-right-outputs">
-                <div className="dub-outputs-row">
-                  <span className="dub-outputs-title-strong">{t('dub.output_options')}</span>
-                  <label>
-                    <input type="checkbox" checked={preserveBg} onChange={e => setPreserveBg(e.target.checked)} /> {t('dub.mix_bg_audio')}
-                  </label>
-                  <label title={t('dub.dual_subs_title')}>
-                    <input type="checkbox" checked={!!dualSubs} onChange={e => setDualSubs(e.target.checked)} /> {t('dub.dual_subs')}
-                  </label>
-                  <label title={t('dub.burn_subs_title')}>
-                    <input type="checkbox" checked={!!burnSubs} onChange={e => setBurnSubs(e.target.checked)} /> {t('dub.burn_subs')}
-                  </label>
-                  <label>
-                    {t('dub.default_track')}
-                    <select className="input-base dub-outputs-default" value={defaultTrack} onChange={e => setDefaultTrack(e.target.value)}>
-                      <option value="original">{t('dub.original_track')}</option>
-                      {dubLangCode && <option value={dubLangCode}>{t('dub.selected_dub', { code: dubLangCode })}</option>}
-                      {dubTracks.filter(tr => tr !== dubLangCode).map(tr => (
-                        <option key={tr} value={tr}>{t('dub.dub_track', { code: tr })}</option>
-                      ))}
-                    </select>
-                  </label>
-                </div>
-                <div className="dub-outputs-row" title="Timing strategy — how the dub reconciles natural-rate TTS with the original timeline.">
-                  <span className="dub-outputs-title-strong">Timing:</span>
-                  <Segmented
-                    value={timingStrategy}
-                    onChange={setTimingStrategy}
-                    items={[
-                      { value: 'concise',       label: 'Concise',        title: 'Translator trims text to fit at natural rate. Overflows surface in the row badge so you can shorten the segment.' },
-                      { value: 'smart_fit',     label: t('dub.timing_smart_fit'), title: t('dub.timing_smart_fit_title') },
-                      { value: 'stretch_video', label: 'Stretch Video',  title: 'Audio plays at natural rate; each segment of the video is stretched (per-segment ffmpeg setpts) to fit. Total video duration grows. Requires a re-encode pass.' },
-                      { value: 'strict_slot',   label: 'Strict slot',    title: 'Legacy: compress audio to fit the original timing. Can sound rushed/chipmunky on high-density target languages.' },
-                    ]}
-                  />
-                </div>
-              </div>
-
-              {dubTranscript && (
-                <div className="dub-transcript-toggle-wrap">
-                  <div className="override-toggle dub-transcript-toggle__inner" onClick={() => setShowTranscript(!showTranscript)}>
-                    <span><FileText size={10} className="dub-inline-icon" /> {t('dub.transcript')}</span>
-                    {showTranscript ? <ChevronUp size={10} /> : <ChevronDown size={10} />}
-                  </div>
-                  {showTranscript && (
-                    <div className="dub-transcript-body">
-                      {dubTranscript}
-                    </div>
-                  )}
-                </div>
-              )}
-
-              {/* Phase 1.3 — Project glossary. Hidden behind a chip until
-                  the user wants it (or terms already exist). */}
-              {dubJobId && !glossaryVisible && (
-                <button
-                  type="button"
-                  className="dub-glossary-chip"
-                  onClick={() => { setGlossaryOpen(true); setGlossaryHidden(false); }}
-                  title={t('dub.glossary_title')}
-                >
-                  {t('dub.glossary_btn', { count: glossaryTermCount })}
-                </button>
-              )}
-              {dubJobId && glossaryVisible && (
-                <div className="dub-glossary-wrap">
-                  <GlossaryPanel
-                    projectId={dubJobId}
-                    sourceLang={dubLangCode && dubLang ? (dubLang.slice(0, 2).toLowerCase() || 'en') : 'en'}
-                    targetLang={dubLangCode}
-                    segments={dubSegments}
-                    onChange={onGlossaryChange}
-                    onClose={() => { setGlossaryHidden(true); setGlossaryOpen(false); }}
-                  />
-                </div>
-              )}
-
-              {/* "Apply Voice to All" row removed 2026-04-21 — redundant
-                  with the CAST strip in the left column, which does the same
-                  thing per-speaker (and handles the multi-speaker case cleanly). */}
-
-              {selectedSegIds.size > 0 && (
-                <div className="dub-bulk-row dub-bulk-row--select">
-                  <span className="dub-bulk-row__label-brand">{t('dub.selected_count', { count: selectedSegIds.size })}</span>
-                  <select className="input-base dub-bulk-select dub-bulk-select--voice"
-                    value="" onChange={(e) => { const v = e.target.value; if (v === '__clear__') bulkApplyToSelected({ profile_id: '' }); else if (v) bulkApplyToSelected({ profile_id: v }); }}>
-                    <option value="">{t('dub.set_voice')}</option>
-                    <option value="__clear__">{t('dub.clear_voice')}</option>
-                    {speakerClones && Object.keys(speakerClones).length > 0 && (
-                      <optgroup label={t('dub.cast')}>
-                        {Object.keys(speakerClones).map(spk => {
-                          const autoId = `auto:${(spk || '').toLowerCase().replace(/\s+/g, '_')}`;
-                          return <option key={autoId} value={autoId}>🎤 {spk}</option>;
-                        })}
-                      </optgroup>
-                    )}
-                    {profiles.filter(p => !p.instruct).length > 0 && (
-                      <optgroup label={t('dub.clone_profiles')}>
-                        {profiles.filter(p => !p.instruct).map(p => <option key={p.id} value={p.id}>{p.name}</option>)}
-                      </optgroup>
-                    )}
-                    {profiles.filter(p => !!p.instruct).length > 0 && (
-                      <optgroup label={t('dub.design_presets')}>
-                        {profiles.filter(p => !!p.instruct).map(p => <option key={p.id} value={p.id}>{p.name}</option>)}
-                      </optgroup>
-                    )}
-                  </select>
-                  <select className="input-base dub-bulk-select dub-bulk-select--lang"
-                    value="" onChange={(e) => { if (e.target.value === '__def__') bulkApplyToSelected({ target_lang: null }); else if (e.target.value) bulkApplyToSelected({ target_lang: e.target.value }); }}>
-                    <option value="">{t('dub.set_lang')}</option>
-                    <option value="__def__">{t('dub.default_lang')}</option>
-                    {LANG_CODES.map(lc => <option key={lc.code} value={lc.code}>{lc.code.toUpperCase()}</option>)}
-                  </select>
-                  <Button variant="danger" size="sm" onClick={bulkDeleteSelected}>{t('dub.delete_selected')}</Button>
-                  <Button variant="ghost"  size="sm" onClick={clearSegSelection} className="dub-bulk-row__clear">{t('dub.clear_selection')}</Button>
-                </div>
-              )}
-
-              {showCheckpoint && (
-                <CheckpointBanner
-                  stage={checkpointStage}
-                  count={dubSegments.length}
-                  onContinue={checkpointStage === 'done' ? null : onCheckpointContinue}
-                  onDismiss={onCheckpointDismiss}
-                  continueLoading={isTranslating}
-                />
-              )}
-
-              <Suspense fallback={<LazyFallback />}>
-                <DubSegmentTable
-                  segments={dubSegments}
-                  profiles={profiles}
-                  speakerClones={speakerClones}
-                  dubStep={dubStep}
-                  dubProgress={dubProgress}
-                  previewLoadingId={segmentPreviewLoading}
-                  selectedIds={selectedSegIds}
-                  onSelect={toggleSegSelect}
-                  onSelectAll={selectAllSegs}
-                  onClearSelection={clearSegSelection}
-                  onEditField={segmentEditField}
-                  onDelete={segmentDelete}
-                  onRestore={segmentRestoreOriginal}
-                  onPreview={handleSegmentPreview}
-                  onDirect={onDirectSegment}
-                  onSplit={segmentSplit}
-                  onMerge={segmentMerge}
-                  onSeek={seekWaveform}
-                  timelineSelectedId={timelineSelSegId}
-                />
-              </Suspense>
-            </div>
+            <DubLeftColumn hasDubbedTrack={hasDubbedTrack} t={t} previewMode={previewMode} setPreviewMode={setPreviewMode} dubTracks={dubTracks} videoSrc={videoSrc} waveformRef={waveformRef} dubJobId={dubJobId} dubSegments={dubSegments} timelineOnsets={timelineOnsets} timelineSelSegId={timelineSelSegId} setTimelineSelSegId={setTimelineSelSegId} incrementalPlan={incrementalPlan} segmentMoveResize={segmentMoveResize} segmentDelete={segmentDelete} onTimelinePreviewSegment={onTimelinePreviewSegment} dubStep={dubStep} dubProgress={dubProgress} fmtDur={fmtDur} genElapsed={genElapsed} genRemaining={genRemaining} speakerClones={speakerClones} setDubSegments={setDubSegments} profiles={profiles} settingsOpen={settingsOpen} setSettingsOpen={setSettingsOpen} dubLang={dubLang} dubLangCode={dubLangCode} translateQuality={translateQuality} activeEngineUnavailable={activeEngineUnavailable} translateProvider={translateProvider} dubInstruct={dubInstruct} setDubInstruct={setDubInstruct} handleTranslateAll={handleTranslateAll} isTranslating={isTranslating} hasAnyTranslation={hasAnyTranslation} handleCleanupSegments={handleCleanupSegments} setDubLang={setDubLang} setDubLangCode={setDubLangCode} dubDialect={dubDialect} setDubDialect={setDubDialect} i18n={i18n} enginesSandboxed={enginesSandboxed} handleInstallEngine={handleInstallEngine} engineInstalling={engineInstalling} activeEngineEntry={activeEngineEntry} engines={engines} setTranslateProvider={setTranslateProvider} setTranslateQuality={setTranslateQuality} llmEndpoint={llmEndpoint} multiLangMode={multiLangMode} setMultiLangMode={setMultiLangMode} multiLangs={multiLangs} setMultiLangs={setMultiLangs} editSegments={editSegments} />
+            <DubRightColumn t={t} preserveBg={preserveBg} setPreserveBg={setPreserveBg} dualSubs={dualSubs} setDualSubs={setDualSubs} burnSubs={burnSubs} setBurnSubs={setBurnSubs} defaultTrack={defaultTrack} setDefaultTrack={setDefaultTrack} dubLangCode={dubLangCode} dubTracks={dubTracks} timingStrategy={timingStrategy} setTimingStrategy={setTimingStrategy} dubTranscript={dubTranscript} showTranscript={showTranscript} setShowTranscript={setShowTranscript} dubJobId={dubJobId} glossaryVisible={glossaryVisible} setGlossaryOpen={setGlossaryOpen} setGlossaryHidden={setGlossaryHidden} glossaryTermCount={glossaryTermCount} dubLang={dubLang} dubSegments={dubSegments} onGlossaryChange={onGlossaryChange} selectedSegIds={selectedSegIds} bulkApplyToSelected={bulkApplyToSelected} speakerClones={speakerClones} profiles={profiles} clearSegSelection={clearSegSelection} bulkDeleteSelected={bulkDeleteSelected} showCheckpoint={showCheckpoint} checkpointStage={checkpointStage} onCheckpointContinue={onCheckpointContinue} onCheckpointDismiss={onCheckpointDismiss} isTranslating={isTranslating} segmentPreviewLoading={segmentPreviewLoading} toggleSegSelect={toggleSegSelect} selectAllSegs={selectAllSegs} segmentEditField={segmentEditField} segmentDelete={segmentDelete} segmentRestoreOriginal={segmentRestoreOriginal} handleSegmentPreview={handleSegmentPreview} onDirectSegment={onDirectSegment} segmentSplit={segmentSplit} segmentMerge={segmentMerge} seekWaveform={seekWaveform} timelineSelSegId={timelineSelSegId} dubStep={dubStep} dubProgress={dubProgress} />
           </div>
-
-          {/* Actions footer */}
-          <div className="studio-panel dub-footer-panel">
-            {dubStep === 'done' && (
-              <div className="dub-footer-banner">
-                <Badge tone="success">
-                  <Check size={11} /> {t('dub.tracks_done', { tracks: dubTracks.join(', ') })}
-                </Badge>
-                {incrementalPlan && incrementalPlan.stale?.length > 0 && (
-                  <Badge tone="warn" className="dub-footer-banner__badge-gap">
-                    {t('dub.segments_changed', { count: incrementalPlan.stale.length })}
-                  </Badge>
-                )}
-                {incrementalPlan && incrementalPlan.stale?.length === 0 && incrementalPlan.fresh?.length > 0 && (
-                  <Badge tone="neutral" className="dub-footer-banner__badge-gap">
-                    {t('dub.all_up_to_date', { count: incrementalPlan.fresh.length })}
-                  </Badge>
-                )}
-              </div>
-            )}
-            {dubError && (
-              <div className="dub-footer-banner">
-                <Badge tone="danger">
-                  <AlertCircle size={11} /> {dubError}
-                </Badge>
-                <DubFailureNotice failure={dubFailure} />
-              </div>
-            )}
-            {/* Output options + Timing moved to the top of the right (transcript) section. */}
-            {dubTracks.length > 0 && (
-              <div className="dub-tracks-row">
-                <span className="dub-tracks-row__title">{t('dub.export_tracks')}</span>
-                <label className={exportTracks['original'] !== false ? 'is-on' : 'is-off'}>
-                  <input type="checkbox" checked={exportTracks['original'] !== false} onChange={e => setExportTracks(prev => ({ ...prev, original: e.target.checked }))} />
-                  <span>{t('dub.original_track')}</span>
-                </label>
-                {dubTracks.map(t => (
-                  <label key={t} className={exportTracks[t] !== false ? 'is-on is-success' : 'is-off'}>
-                    <input type="checkbox" checked={exportTracks[t] !== false} onChange={e => setExportTracks(prev => ({ ...prev, [t]: e.target.checked }))} />
-                    <span className="code">{t}</span>
-                  </label>
-                ))}
-              </div>
-            )}
-            {(() => {
-              // Pre-generation compression warning. Predicted by the
-              // translate response (see services/speech_rate.rate_ratio
-              // + dub_translate._maybe_cinematic), populated whenever
-              // segments carry a slot_seconds and translated text.
-              // Surfaces here so the user can act (re-translate in
-              // Cinematic, edit text, allow longer slots) before
-              // committing to a full Generate Dub run.
-              const hot = dubSegments.filter(s => (s.rate_ratio || 0) > 1.3);
-              if (hot.length === 0 || !dubSegments.length) return null;
-              const pctHot = Math.round((hot.length / dubSegments.length) * 100);
-              if (pctHot < 10) return null;
-              const worst = hot.reduce((a, b) => (a.rate_ratio > b.rate_ratio ? a : b));
-              return (
-                <div className="dub-compression-warn" role="status">
-                  <span className="dub-compression-warn__icon">⚠</span>
-                  <span className="dub-compression-warn__body">
-                    <strong>{hot.length} of {dubSegments.length}</strong> segments need {'>'}1.3× compression
-                    (worst: <span style={{ fontVariantNumeric: 'tabular-nums' }}>{worst.rate_ratio.toFixed(2)}×</span>).
-                    Output will be intelligible (pitch-preserving stretch) but stressed —
-                    {translateQuality === 'fast' ? ' switch to Cinematic and Re-translate' : ' shorten the worst segments'}
-                    {' '}for cleaner audio.
-                  </span>
-                </div>
-              );
-            })()}
-            {/* Generate / Export / Stop actions moved to the header bar (dub-head__primary). */}
-          </div>
+          <DubFooter t={t} dubStep={dubStep} dubTracks={dubTracks} incrementalPlan={incrementalPlan} dubError={dubError} dubFailure={dubFailure} exportTracks={exportTracks} setExportTracks={setExportTracks} dubSegments={dubSegments} translateQuality={translateQuality} />
         </div>
       )}
